Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

Base.finalize

Base.finalize

Base.finalize

Base.finalize

MUMPS.associate_matrix!

MUMPS.associate_matrix!

MUMPS.associate_matrix!

MUMPS.associate_matrix!

MUMPS.associate_rhs!

MUMPS.associate_rhs!

MUMPS.associate_rhs!

MUMPS.associate_rhs!

MUMPS.factorize!

MUMPS.factorize!

MUMPS.factorize!

MUMPS.factorize!

MUMPS.factorize!

MUMPS.factorize!

MUMPS.get_icntl

MUMPS.get_solution

MUMPS.get_solution

MUMPS.get_solution

MUMPS.get_solution

MUMPS.solve

MUMPS.solve

MUMPS.solve

MUMPS.solve!

MUMPS.solve!

MUMPS.solve!

MUMPS.solve!

​

Base.finalize — Method

Terminate a Mumps instance.

Base.finalize — Method

Terminate a Mumps instance.

Base.finalize — Method

Terminate a Mumps instance.

Base.finalize — Method

Terminate a Mumps instance.

MUMPS.associate_matrix! — Method

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

MUMPS.associate_matrix! — Method

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

MUMPS.associate_matrix! — Method

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

MUMPS.associate_matrix! — Method

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

MUMPS.associate_rhs! — Method

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

MUMPS.associate_rhs! — Method

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

MUMPS.associate_rhs! — Method

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

MUMPS.associate_rhs! — Method

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

MUMPS.factorize! — Method

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

MUMPS.factorize! — Method

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

MUMPS.factorize! — Method

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

MUMPS.factorize! — Method

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

MUMPS.factorize! — Method

Combined associate_matrix / factorize. Presume that A is available on all nodes.

MUMPS.factorize! — Method

Combined associate_matrix / factorize. Presume that A is available on all nodes.

MUMPS.get_icntl — Method

Obtain an array of integer control parameters.

MUMPS.get_solution — Method

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.

MUMPS.get_solution — Method

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.

MUMPS.get_solution — Method

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.

MUMPS.get_solution — Method

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.

MUMPS.solve! — Method

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

MUMPS.solve! — Method

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

MUMPS.solve! — Method

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

MUMPS.solve! — Method

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

MUMPS.solve — Method

Combined initialize / analyze / factorize / solve. Presume that A and rhs are available on all nodes. The optional keyword argument sym indicates the symmetry of A. The solution is retrieved and returned.

MUMPS.solve — Method

Combined analyze / factorize / solve. Presume that A and rhs are available on all nodes. The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is retrieved and returned.

MUMPS.solve — Method

Combined associate_rhs / solve. Presume that rhs is available on all nodes. The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is retrieved and returned.