API

API

Constants

Types

Utilities

Constants

MUMPS.default_icntl — Constant

Default integer parameters.

MUMPS.default_cntl32 — Constant

Default single precision real parameters

MUMPS.default_cntl64 — Constant

Default double precision real parameters

MUMPS.mumps_unsymmetric — Constant

Constant indicating that a general unsymmetric matrix will be analyzed and factorized

MUMPS.mumps_definite — Constant

Constant indicating that a symmetric definite matrix will be analyzed and factorized

MUMPS.mumps_symmetric — Constant

Constant indicating that a general symmetric matrix will be analyzed and factorized

Types

MUMPS.MUMPSException — Type

Exception type raised in case of error.

MUMPS.Mumps — Type

Abstract type representing a factorization with MUMPS. All constructor arguments are optional. By default a general unsymmetric matrix will be analyzed/factorized with default integer and real parameters

Utilities

MUMPS.get_icntl — Function

Obtain an array of integer control parameters.

Base.finalize — Function

Terminate a Mumps instance.

Terminate a Mumps instance.

Terminate a Mumps instance.

Terminate a Mumps instance.

MUMPS.factorize! — Function

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

Combined associate_matrix / factorize. Presume that A is available on all nodes.

Combined associate_matrix / factorize. Presume that A is available on all nodes.

MUMPS.solve! — Function

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

MUMPS.solve — Function

Combined associate_rhs / solve. Presume that rhs is available on all nodes. The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is retrieved and returned.

Combined analyze / factorize / solve. Presume that A and rhs are available on all nodes. The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is retrieved and returned.

Combined initialize / analyze / factorize / solve. Presume that A and rhs are available on all nodes. The optional keyword argument sym indicates the symmetry of A. The solution is retrieved and returned.

MUMPS.associate_matrix! — Function

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

Register the matrix A with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

MUMPS.associate_rhs! — Function

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

Register the right-hand side(s) rhs with the Mumps object mumps. This function makes it possible to define the right- -hand side(s) on the host only. If the right-hand side(s) are defined on all nodes, there is no need to use this function.

MUMPS.get_solution — Function

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.