API

API

Constants

Types

Utilities

Constants

MUMPS.default_icntl — Constant

Default integer parameters.

MUMPS.default_cntl32 — Constant

Default single precision real parameters

MUMPS.default_cntl64 — Constant

Default double precision real parameters

MUMPS.mumps_unsymmetric — Constant

const mumps_unsymmetric

Constant indicating that a general unsymmetric matrix will be analyzed and factorized

MUMPS.mumps_definite — Constant

const mumps_definite

Constant indicating that a symmetric definite matrix will be analyzed and factorized

MUMPS.mumps_symmetric — Constant

const mumps_symmetric

Constant indicating that a general symmetric matrix will be analyzed and factorized

Types

MUMPS.MUMPSException — Type

mutable struct MUMPSException

Exception type raised in case of error.

Missing docstring.

Missing docstring for Mumps. Check Documenter's build log for details.

Utilities

MUMPS.get_icntl — Function

get_icntl(;det=false,verbose=false,ooc=false,itref=0)

Obtain an array of integer control parameters.

Base.finalize — Function

finalize(mumps)

Terminate a Mumps instance.

MUMPS.factorize! — Function

factorize!(mumps)

Factorize the matrix registered with the Mumps instance. The matrix must have been previously registered with associate_matrix(). After the factorization, the determinant, if requested, is stored in mumps.det. The MUMPS error code is stored in mumps.err.

factorize!(mumps,A)

Combined associate_matrix / factorize. Presume that A is available on all nodes.

MUMPS.solve! — Function

solve!(mumps;transposed=false)

Solve the system registered with the Mumps object mumps. The matrix and right-hand side(s) must have been previously registered with associate_matrix() and associate_rhs(). The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is stored internally and must be retrieved with get_solution().

MUMPS.solve — Function

solve(mumps, rhs; transposed=false)

Combined associate_rhs / solve. Presume that rhs is available on all nodes. The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is retrieved and returned.

solve(mumps,A,rhs; transposed=false)

Combined analyze / factorize / solve. Presume that A and rhs are available on all nodes. The optional keyword argument transposed indicates whether the user wants to solve the forward or transposed system. The solution is retrieved and returned.

solve(A,rhs;sym=mumps_unsymmetric)

Combined initialize / analyze / factorize / solve. Presume that A and rhs are available on all nodes. The optional keyword argument sym indicates the symmetry of A. The solution is retrieved and returned.

MUMPS.associate_matrix! — Function

associate_matrix!(mumps,A)

Register the square matrix A to a mumps object. It internally converts A to be consistent with the ICNTL[5] setting.

If needed, it tries to convert element type of A to be consistent with type of mumps, throwing a warning in this case.

See also: associate_rhs!

associate_matrix!(mumps, n, irow, jcol, vals)

Register the sparse matrix given in coordinate format with the Mumps object mumps. This function makes it possible to define the matrix on the host only. If the matrix is defined on all nodes, there is no need to use this function.

MUMPS.associate_rhs! — Function

associate_rhs!(mumps, rhs)

Register a dense or sparse RHS matrix or vector rhs to a mumps object. It internally converts rhs to be consistent with the ICNTL[20] setting, and additionally allocates mumps.rhs according to the ICNTL[21] setting.

If needed, it tries to convert element type of rhs to be consistent with type of mumps, throwing a warning in this case.

See also: associate_matrix!

MUMPS.get_solution — Function

get_solution(mumps) -> x

Retrieve the solution of the system solved by solve(). This function makes it possible to ask MUMPS to assemble the final solution on the host only, and to retrieve it there.